refactor: split identity FFI module into organized submodules
- Split identity.rs into mod.rs with organized submodules
- Created queries/ submodule with individual files for each query operation
- Organized functions into logical modules:
  - helpers.rs: utility functions and converters
  - create.rs: identity creation
  - topup.rs: top-up operations
  - put.rs: put-to-platform operations
  - transfer.rs: credit transfer operations
  - withdraw.rs: withdrawal operations
  - info.rs: identity info and lifecycle
  - names.rs: name registration
  - queries/: all query operations split into separate files
