+++ packages/ios-sdk-ffi/src/identity/create.rs
@@ -1,0 +1,21 @@
+//! Identity creation operations
+
+use crate::types::SDKHandle;
+use crate::{IOSSDKError, IOSSDKErrorCode, IOSSDKResult};
+
+/// Create a new identity
+#[no_mangle]
+pub unsafe extern "C" fn ios_sdk_identity_create(sdk_handle: *mut SDKHandle) -> IOSSDKResult {
+    if sdk_handle.is_null() {
+        return IOSSDKResult::error(IOSSDKError::new(
+            IOSSDKErrorCode::InvalidParameter,
+            "SDK handle is null".to_string(),
+        ));
+    }
+
+    // TODO: Implement identity creation once the SDK API is available
+    IOSSDKResult::error(IOSSDKError::new(
+        IOSSDKErrorCode::NotImplemented,
+        "Identity creation not yet implemented".to_string(),
+    ))
+}
