+++ packages/ios-sdk-ffi/src/identity/queries/fetch.rs
@@ -1,0 +1,72 @@
+//! Identity fetch operations
+
+use dash_sdk::dpp::platform_value::string_encoding::Encoding;
+use dash_sdk::dpp::prelude::{Identifier, Identity};
+use dash_sdk::platform::Fetch;
+use std::ffi::CStr;
+use std::os::raw::c_char;
+
+use crate::sdk::SDKWrapper;
+use crate::types::{IOSSDKResultDataType, IdentityHandle, SDKHandle};
+use crate::{FFIError, IOSSDKError, IOSSDKErrorCode, IOSSDKResult};
+
+/// Fetch an identity by ID
+#[no_mangle]
+pub unsafe extern "C" fn ios_sdk_identity_fetch(
+    sdk_handle: *const SDKHandle,
+    identity_id: *const c_char,
+) -> IOSSDKResult {
+    if sdk_handle.is_null() {
+        return IOSSDKResult::error(IOSSDKError::new(
+            IOSSDKErrorCode::InvalidParameter,
+            "SDK handle is null".to_string(),
+        ));
+    }
+
+    if identity_id.is_null() {
+        return IOSSDKResult::error(IOSSDKError::new(
+            IOSSDKErrorCode::InvalidParameter,
+            "Identity ID is null".to_string(),
+        ));
+    }
+
+    let wrapper = &*(sdk_handle as *const SDKWrapper);
+
+    let id_str = match CStr::from_ptr(identity_id).to_str() {
+        Ok(s) => s,
+        Err(e) => {
+            return IOSSDKResult::error(FFIError::from(e).into());
+        }
+    };
+
+    let id = match Identifier::from_string(id_str, Encoding::Base58) {
+        Ok(id) => id,
+        Err(e) => {
+            return IOSSDKResult::error(IOSSDKError::new(
+                IOSSDKErrorCode::InvalidParameter,
+                format!("Invalid identity ID: {}", e),
+            ));
+        }
+    };
+
+    let result = wrapper.runtime.block_on(async {
+        Identity::fetch(&wrapper.sdk, id)
+            .await
+            .map_err(FFIError::from)
+    });
+
+    match result {
+        Ok(Some(identity)) => {
+            let handle = Box::into_raw(Box::new(identity)) as *mut IdentityHandle;
+            IOSSDKResult::success_handle(
+                handle as *mut std::os::raw::c_void,
+                IOSSDKResultDataType::IdentityHandle,
+            )
+        }
+        Ok(None) => IOSSDKResult::error(IOSSDKError::new(
+            IOSSDKErrorCode::NotFound,
+            "Identity not found".to_string(),
+        )),
+        Err(e) => IOSSDKResult::error(e.into()),
+    }
+}
