+++ packages/ios-sdk-ffi/src/identity/queries/mod.rs
@@ -1,0 +1,12 @@
+//! Identity query operations
+
+pub mod balance;
+pub mod fetch;
+pub mod public_keys;
+pub mod resolve;
+
+// Re-export all public functions for convenient access
+pub use balance::ios_sdk_identity_fetch_balance;
+pub use fetch::ios_sdk_identity_fetch;
+pub use public_keys::ios_sdk_identity_fetch_public_keys;
+pub use resolve::ios_sdk_identity_resolve_name;
